creation route sauce ok

+++ controllers/sauce.js
@@ -1,15 +1,15 @@
 const Sauce = require ('../models/Sauce');
+const fs = require('fs');
 
-exports.getAllThings = (req, res, next) => {
-    Sauce.find()
-        .then( sauce => res.status(200).json(sauce))
-        .catch(error => res.status(400).json({ error }));
-        console.log("=======ppp==========");
-};
+
 exports.createThing = (req, res, next) => {
-    delete req.body._id;
+    const sauceObject = JSON.parse(req.body.sauce);
+    delete sauceObject._id;
+    delete sauceObject._userId;
     const sauce = new Sauce({
-        ...req.body
+        ...sauceObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
     sauce.save()
         .then(() => res.status(201).json({message: 'Sauce enregistrée !'}))
@@ -17,16 +17,45 @@ exports.createThing = (req, res, next) => {
 };
 
 exports.modifyThing = (req, res, next) => {
-    Sauce.updateOne({ _id: req.params.id }, { ...req.body, _id:req.params.id})
-        .then(() => res.status(200).json({ message: 'Sauce modifiée !'}))
-        .catch(error => res.status(400).json({ error}));
-};
+    const sauceObject = req.file ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+    
+    delete sauceObject._userId;
+    Sauce.findOne({_id: req.params.id})
+        .then((sauce) => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({ message : 'Non autorisé !'});
+            } else {
+                Sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
+                .then(() => res.status(200).json({message : 'Objet modifié!'}))
+                .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
+ };
 
-exports.deleteThing = ( req, res, next) => {
-    Sauce.deleteOne({ _id: req.params.id})
-    .then(() => res.status(200).json({message: 'Sauce supprimée !'}))
-    .catch(error => res.status(404).json({ error }));
-};
+ exports.deleteThing = (req, res, next) => {
+    Sauce.findOne({ _id: req.params.id})
+        .then(sauce => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({message: 'Not authorized'});
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch( error => {
+            res.status(500).json({ error });
+        });
+ };
 
 exports.getOneThing = ( req, res, next) => {
     Sauce.findOne({ _id: req.params.id})
@@ -34,3 +63,8 @@ exports.getOneThing = ( req, res, next) => {
         .catch (error => res.status(404).json({ error }));
 };
 
+exports.getAllThings = (req, res, next) => {
+    Sauce.find()
+        .then( sauce => res.status(200).json(sauce))
+        .catch(error => res.status(400).json({ error }));
+};
